+++ csrc/phi/moe/utils/moe_align_block_size_kernels.cu
@@ -103,8 +103,8 @@ __global__ void moe_align_block_size_kernel(scalar_t *__restrict__ topk_ids,
 
 void moe_align_block_size(
     torch::Tensor topk_ids,
-    int num_experts,
-    int block_size,
+    int64_t num_experts,
+    int64_t block_size,
     torch::Tensor sorted_token_ids,
     torch::Tensor experts_ids,
     torch::Tensor num_tokens_post_pad,
